@@ -2,7 +2,7 @@ package com.shubham.iitg.blogapp;
 
 
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
+import androidx.core.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
